mushrooms/requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.4.0 